+++ libs/maps/src/opengl/CAngularObservationMesh.cpp
@@ -181,6 +181,9 @@ void CAngularObservationMesh::onUpdateBuffers_Wireframe()
 {
 	auto& vbd = CRenderizableShaderWireFrame::m_vertex_buffer_data;
 	auto& cbd = CRenderizableShaderWireFrame::m_color_buffer_data;
+	std::unique_lock<std::shared_mutex> wfWriteLock(
+		CRenderizableShaderWireFrame::m_wireframeMtx.data);
+
 	vbd.clear();
 	cbd.clear();
 
@@ -198,6 +201,8 @@ void CAngularObservationMesh::onUpdateBuffers_Wireframe()
 
 void CAngularObservationMesh::onUpdateBuffers_Triangles()
 {
+	std::unique_lock<std::shared_mutex> trisWriteLock(
+		CRenderizableShaderTriangles::m_trianglesMtx.data);
 	auto& tris = CRenderizableShaderTriangles::m_triangles;
 
 	tris = this->triangles;