+++ libs/opengl/include/mrpt/opengl/CRenderizable.h
@@ -8,6 +8,8 @@
    +------------------------------------------------------------------------+ */
 #pragma once
 
+#include <mrpt/containers/NonCopiableData.h>
+#include <mrpt/containers/PerThreadDataHolder.h>
 #include <mrpt/containers/yaml_frwd.h>
 #include <mrpt/img/TColor.h>
 #include <mrpt/math/TBoundingBox.h>
@@ -24,18 +26,19 @@
 #include <mrpt/typemeta/TEnumType.h>
 
 #include <deque>
+#include <shared_mutex>
 
 namespace mrpt::opengl
 {
 /** The base class of 3D objects that can be directly rendered through OpenGL.
  *  In this class there are a set of common properties to all 3D objects,
  *mainly:
- * - A name (m_name): A name that can be optionally asigned to objects for
+ * - Its SE(3) pose (x,y,z,yaw,pitch,roll), relative to the parent object,
+ * or the global frame of reference for root objects (inserted into a
+ *mrpt::opengl::COpenGLScene).
+ * - A name: A name that can be optionally asigned to objects for
  *easing its reference.
- * - 6D coordinates (x,y,z,yaw,pitch,roll), relative to the "current"
- *reference framework. By default, any object is referenced to global scene
- *coordinates.
- * - A RGB color: This field will be used in simple elements (points,
+ * - A RGBA color: This field will be used in simple elements (points,
  *lines, text,...) but is ignored in more complex objects that carry their own
  *color information (triangle sets,...)
  *
@@ -310,21 +313,28 @@ class CRenderizable : public mrpt::serialization::CSerializable
 	void updateBuffers() const
 	{
 		renderUpdateBuffers();
-		const_cast<CRenderizable&>(*this).m_outdatedBuffers = false;
+		std::unique_lock<std::shared_mutex> lckWrite(m_stateMtx.data);
+		const_cast<CRenderizable&>(*this).m_state.get().outdatedBuffers = false;
 	}
 
 	/** Call to enable calling renderUpdateBuffers() before the next
 	 * render() rendering iteration. */
 	void notifyChange() const
 	{
-		const_cast<CRenderizable&>(*this).m_outdatedBuffers = true;
+		std::unique_lock<std::shared_mutex> lckWrite(m_stateMtx.data);
+		const_cast<CRenderizable&>(*this).m_state.run_on_all(
+			[](auto& state) { state.outdatedBuffers = true; });
 	}
 
 	/** Returns whether notifyChange() has been invoked since the last call
 	 * to renderUpdateBuffers(), meaning the latter needs to be called again
 	 * before rendering.
 	 */
-	bool hasToUpdateBuffers() const { return m_outdatedBuffers; }
+	bool hasToUpdateBuffers() const
+	{
+		std::shared_lock<std::shared_mutex> lckWrite(m_stateMtx.data);
+		return m_state.get().outdatedBuffers;
+	}
 
 	/** Simulation of ray-trace, given a pose. Returns true if the ray
 	 * effectively collisions with the object (returning the distance to the
@@ -377,7 +387,13 @@ class CRenderizable : public mrpt::serialization::CSerializable
 	void writeToStreamRender(mrpt::serialization::CArchive& out) const;
 	void readFromStreamRender(mrpt::serialization::CArchive& in);
 
-	bool m_outdatedBuffers = true;
+	struct State
+	{
+		bool outdatedBuffers = true;
+	};
+	mrpt::containers::PerThreadDataHolder<State> m_state;
+	mutable mrpt::containers::NonCopiableData<std::shared_mutex> m_stateMtx;
+
 	mrpt::math::TPoint3Df m_representativePoint{0, 0, 0};
 
 	/** Optional pointer to a mrpt::opengl::CText */