+++ libs/opengl/src/CBox.cpp
@@ -54,6 +54,9 @@ void CBox::onUpdateBuffers_Wireframe()
 {
 	auto& vbd = CRenderizableShaderWireFrame::m_vertex_buffer_data;
 	auto& cbd = CRenderizableShaderWireFrame::m_color_buffer_data;
+	std::unique_lock<std::shared_mutex> wfWriteLock(
+		CRenderizableShaderWireFrame::m_wireframeMtx.data);
+
 	vbd.clear();
 
 	const std::array<mrpt::math::TPoint3D, 2> corner = {
@@ -84,7 +87,10 @@ void CBox::onUpdateBuffers_Wireframe()
 }
 void CBox::onUpdateBuffers_Triangles()
 {
+	std::unique_lock<std::shared_mutex> trisWriteLock(
+		CRenderizableShaderTriangles::m_trianglesMtx.data);
 	auto& tris = CRenderizableShaderTriangles::m_triangles;
+
 	tris.clear();
 
 	const auto &c0 = m_corner_min, &c1 = m_corner_max;