+++ libs/opengl/src/CCylinder.cpp
@@ -24,7 +24,10 @@ IMPLEMENTS_SERIALIZABLE(CCylinder, CRenderizable, mrpt::opengl)
 
 void CCylinder::onUpdateBuffers_Triangles()
 {
+	std::unique_lock<std::shared_mutex> trisWriteLock(
+		CRenderizableShaderTriangles::m_trianglesMtx.data);
 	auto& tris = CRenderizableShaderTriangles::m_triangles;
+
 	tris.clear();
 
 	// precomputed table: