+++ libs/opengl/src/CDisk.cpp
@@ -23,7 +23,10 @@ void CDisk::onUpdateBuffers_Triangles()
 {
 	using mrpt::math::TPoint3Df;
 
+	std::unique_lock<std::shared_mutex> trisWriteLock(
+		CRenderizableShaderTriangles::m_trianglesMtx.data);
 	auto& tris = CRenderizableShaderTriangles::m_triangles;
+
 	tris.clear();
 
 	// precomputed table: