+++ libs/opengl/src/CMeshFast.cpp
@@ -65,6 +65,9 @@ void CMeshFast::onUpdateBuffers_Points()
 
 	auto& vbd = CRenderizableShaderPoints::m_vertex_buffer_data;
 	auto& cbd = CRenderizableShaderPoints::m_color_buffer_data;
+	std::unique_lock<std::shared_mutex> wfWriteLock(
+		CRenderizableShaderPoints::m_pointsMtx.data);
+
 	vbd.clear();
 	cbd.clear();