+++ libs/opengl/src/CPolyhedron.cpp
@@ -2218,6 +2218,9 @@ void CPolyhedron::onUpdateBuffers_Wireframe()
 {
 	auto& vbd = CRenderizableShaderWireFrame::m_vertex_buffer_data;
 	auto& cbd = CRenderizableShaderWireFrame::m_color_buffer_data;
+	std::unique_lock<std::shared_mutex> wfWriteLock(
+		CRenderizableShaderWireFrame::m_wireframeMtx.data);
+
 	vbd.clear();
 
 	for (const auto& edge : m_Edges)
@@ -2231,7 +2234,10 @@ void CPolyhedron::onUpdateBuffers_Wireframe()
 
 void CPolyhedron::onUpdateBuffers_Triangles()
 {
+	std::unique_lock<std::shared_mutex> trisWriteLock(
+		CRenderizableShaderTriangles::m_trianglesMtx.data);
 	auto& tris = CRenderizableShaderTriangles::m_triangles;
+
 	tris.clear();
 
 	for (const auto& face : m_Faces)