+++ doc/source/doxygen-docs/changelog.md
@@ -2,6 +2,10 @@
 
 # Version 2.5.6: UNRELEASED
 - Changes in libraries:
+  - \ref mrpt_containers_grp
+    - New template class mrpt::containers::PerThreadDataHolder
+  - \ref mrpt_opengl_grp
+    - All opengl shaders, objects, and renderers are now multithread capable, including rendering several FBOs in different threads.
   - \ref mrpt_system_grp
     - These functions are now thread-safe if built in a system with the `localtime_r()` variant of `localtime()`:
       - mrpt::system::timestampToParts()